requirements.txt
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
google-auth==2.28.1
google-auth-oauthlib==1.2.0
gspread==6.0.2
idna==3.6
logging==0.4.9.6
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.1
patool==2.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
StrEnum==0.4.15
tzdata==2024.1
urllib3==2.2.1
